fix: upstream image response failed in dev mode
.next/cache/images is not created on initial rendering of run dev
After a refresh or page routing it works fine
