+++ templates/next-pages/package.json
@@ -9,12 +9,12 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@lit-protocol/auth-helpers": "^3.0.32",
-    "@lit-protocol/constants": "^3.0.32",
-    "@lit-protocol/contracts-sdk": "^3.0.32",
-    "@lit-protocol/lit-auth-client": "^3.0.32",
-    "@lit-protocol/pkp-ethers": "^3.0.32",
-    "@lit-protocol/types": "^3.0.32",
+    "@lit-protocol/auth-helpers": "3.0.32",
+    "@lit-protocol/constants": "3.0.32",
+    "@lit-protocol/contracts-sdk": "3.0.32",
+    "@lit-protocol/lit-auth-client": "3.0.32",
+    "@lit-protocol/pkp-ethers": "3.0.32",
+    "@lit-protocol/types": "3.0.32",
     "@radix-ui/react-radio-group": "^1.1.3",
     "@stytch/nextjs": "^17.0.0",
     "@stytch/vanilla-js": "^4.3.2",
@@ -27,6 +27,7 @@
     "prettier": "^3.2.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "sharp": "^0.33.2",
     "wagmi": "^0.12.19"
   },
   "devDependencies": {
