+++ templates/next-pages/src/pages/login.tsx
@@ -48,7 +48,7 @@ export default function LoginView() {
   }
 
   function goToSignUp() {
-    router.push('/');
+    router.push('/signup');
   }
 
   useEffect(() => {
